Delete no longer required api directory and move remaining classes to keypoints.ts
@openfish/client negates the need for 90% of the api code, so we only
really need to keep the Keypoint and BoundingBox classes.
