+++ site/src/api.ts
@@ -1,0 +1,3 @@
+import { createClient } from '@openfish/client'
+
+export const client = createClient()
