--- site/src/api/videostream.ts
@@ -1,7 +1,0 @@
-export type VideoStream = {
-  id: number
-  stream_url: string
-  capturesource: string
-  startTime: string
-  endTime: string
-}
