+++ site/src/utils/keypoints.ts
@@ -1,4 +1,49 @@
-import { BoundingBox, type Keypoint } from '../api/annotation'
+import 'reflect-metadata'
+
+import { Transform, Type } from 'class-transformer'
+import { formatVideoTime, parseVideoTime } from '../utils/datetime'
+
+export class Keypoint {
+  @Transform(({ value }) => parseVideoTime(value), { toClassOnly: true })
+  @Transform(({ value }) => formatVideoTime(value, true), { toPlainOnly: true })
+  time: number
+
+  @Type(() => BoundingBox)
+  box: BoundingBox
+
+  constructor(time: number, box: BoundingBox) {
+    this.time = time
+    this.box = box
+  }
+}
+
+export class BoundingBox {
+  x1: number
+  y1: number
+  x2: number
+  y2: number
+
+  constructor(x1: number, y1: number, x2: number, y2: number) {
+    this.x1 = x1
+    this.y1 = y1
+    this.x2 = x2
+    this.y2 = y2
+  }
+
+  get w(): number {
+    return Math.abs(this.x1 - this.x2)
+  }
+  get h(): number {
+    return Math.abs(this.y1 - this.y2)
+  }
+
+  get xmin(): number {
+    return Math.min(this.x1, this.x2)
+  }
+  get ymin(): number {
+    return Math.min(this.y1, this.y2)
+  }
+}
 
 // Keypoints must be sorted.
 export function findClosestKeypointPair(
